Generate SignalFromUI struct and peek function
A full representation of the SignalFromUI type on the C side, along with
a function on the Haskell side that can convert back to the Haskell
type.

This requires HaskadesBinding to use hsc2hs, which does the CPP itself
so GHC no longer needs to (though you need to be sure your settings CPP
is set up for hsc2hs).

hsc2hs cannot read C++ at all, so this necessitates a seperate header
file for some things.